pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "model-maker"
version = "1.0.0"
description = "Offline AI assistant for emergency, medical, and survival situations"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Maitham", email = "maitham@flexsell.com" },
]
keywords = [
    "offline-ai", "llm", "medical", "emergency", "survival",
    "arabic", "multilingual", "gguf", "llama-cpp", "fastapi",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "llama-cpp-python>=0.2.0",
    "pydantic>=2.0.0",
    "onnxruntime>=1.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.24",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
training = [
    "mlx>=0.5.0",
    "mlx-lm>=0.5.0",
    "transformers>=4.36.0",
    "datasets>=2.16.0",
]

[project.urls]
Homepage = "https://github.com/Maitham1988/model-maker"
Documentation = "https://github.com/Maitham1988/model-maker/tree/main/docs"
Repository = "https://github.com/Maitham1988/model-maker"
Issues = "https://github.com/Maitham1988/model-maker/issues"
Changelog = "https://github.com/Maitham1988/model-maker/blob/main/CHANGELOG.md"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "llm: marks tests that require a loaded LLM model",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
